Feat/paginacao

+++ src/main/java/me/gfumagali/btgpedidos/model/documents/OrderTotalValue.java
@@ -5,11 +5,13 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @Document(collection = "orderTotalValue")
 public class OrderTotalValue {
     @Id
     private Long orderCode;
-    private Double totalValue;
+    private BigDecimal totalValue;
 }
